% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_learners.R
\name{available_learners}
\alias{available_learners}
\title{Learners Available for Use}
\usage{
available_learners(outcome_type = c("binomial", "continuous"))
}
\arguments{
\item{outcome_type}{The outcome variable type.}
}
\value{
A \code{data.table} of available learners.
}
\description{
Learners Available for Use
}
\examples{
available_learners("binomial")
}
