% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3superlearner.R
\name{predict.mlr3superlearner}
\alias{predict.mlr3superlearner}
\title{Predict method for \code{mlr3superlearner} object}
\usage{
\method{predict}{mlr3superlearner}(object, newdata)
}
\arguments{
\item{object}{An object returned from \code{mlr3superlearner()}.}

\item{newdata}{A \code{data.frame} to return predictions from.}
}
\value{
Predicted values.
}
\description{
Predict method for \code{mlr3superlearner} object
}
\seealso{
\code{\link{mlr3superlearner}}
}
